023 Done. 회의시간 지나야 투표 가능토록 구현함. 투표시간 지났음에도 투표 안했으면 강제 기권하게 구현함.

+++ Assets/Scripts/GameRuleStore.cs
@@ -407,7 +407,7 @@ public struct GameRuleData
     public bool confirmEjects;
     public int emergencyMeetings;
     public int emergencyMeetingsCooldown;
-    public int meetingsTime;
+    public float meetingsTime;
     public int voteTime;
     public bool anonymousVotes;
     public float moveSpeed;
